separate db into 2 objects for greater efficiency
