+++ tests/test_connection.py
@@ -1,0 +1,15 @@
+import unittest
+from pysqlc import DB
+
+class ConnectTest(unittest.TestCase):
+    """
+    test various features of a dbconfig() object
+    """
+
+    def connectionEnvNoDB(self):
+        db = DB()
+        
+
+
+if __name__ == "__main__":
+    unittest.main()
